+++ src/main/java/net/ow/movie/tmdb/model/movie/TMDBCredits.java
@@ -4,6 +4,8 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import net.ow.movie.tmdb.model.person.TMDBCast;
+import net.ow.movie.tmdb.model.person.TMDBCrew;
 
 @Data
 @NoArgsConstructor
